added ptools router & bumped latest version
