+++ sms_api/api/client/models/ecoli_experiment.py
@@ -12,9 +12,9 @@ from typing import cast, Union
 from typing import Union
 
 if TYPE_CHECKING:
-    from ..models.ecoli_simulation import EcoliSimulation
-    from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
     from ..models.antibiotic_simulation import AntibioticSimulation
+    from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
+    from ..models.ecoli_simulation import EcoliSimulation
 
 
 T = TypeVar("T", bound="EcoliExperiment")
@@ -37,9 +37,9 @@ class EcoliExperiment:
     additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
     def to_dict(self) -> dict[str, Any]:
-        from ..models.ecoli_simulation import EcoliSimulation
-        from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
         from ..models.antibiotic_simulation import AntibioticSimulation
+        from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
+        from ..models.ecoli_simulation import EcoliSimulation
 
         experiment_id = self.experiment_id
 
@@ -70,9 +70,9 @@ class EcoliExperiment:
 
     @classmethod
     def from_dict(cls: type[T], src_dict: Mapping[str, Any]) -> T:
-        from ..models.ecoli_simulation import EcoliSimulation
-        from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
         from ..models.antibiotic_simulation import AntibioticSimulation
+        from ..models.ecoli_experiment_metadata import EcoliExperimentMetadata
+        from ..models.ecoli_simulation import EcoliSimulation
 
         d = dict(src_dict)
         experiment_id = d.pop("experiment_id")
