+++ sms_api/api/client/models/parca_dataset_request.py
@@ -9,8 +9,8 @@ from ..types import UNSET, Unset
 from typing import cast
 
 if TYPE_CHECKING:
-    from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
     from ..models.simulator_version import SimulatorVersion
+    from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
 
 
 T = TypeVar("T", bound="ParcaDatasetRequest")
@@ -29,8 +29,8 @@ class ParcaDatasetRequest:
     additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
     def to_dict(self) -> dict[str, Any]:
-        from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
         from ..models.simulator_version import SimulatorVersion
+        from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
 
         simulator_version = self.simulator_version.to_dict()
 
@@ -47,8 +47,8 @@ class ParcaDatasetRequest:
 
     @classmethod
     def from_dict(cls: type[T], src_dict: Mapping[str, Any]) -> T:
-        from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
         from ..models.simulator_version import SimulatorVersion
+        from ..models.parca_dataset_request_parca_config import ParcaDatasetRequestParcaConfig
 
         d = dict(src_dict)
         simulator_version = SimulatorVersion.from_dict(d.pop("simulator_version"))
