+++ sms_api/api/main.py
@@ -67,7 +67,7 @@ APP_ORIGINS = [
 #     {"url": ServerMode.PORT_FORWARD_DEV, "description": "Local port-forward"},
 # ]
 APP_SERVERS = None
-APP_ROUTERS = ["core"]  # for now, just referencing core
+APP_ROUTERS = ["core", "ptools"]  # for now, just referencing core
 # APP_ROUTERS = ["core", "antibiotic"]
 assets_dir = Path(get_settings().assets_dir)
 ACTIVE_URL = ServerMode.detect(assets_dir / "dev" / "config" / ".dev_env")
