+++ sms_api/api/routers/core.py
@@ -37,8 +37,6 @@ from sms_api.common.gateway.io import get_zip_buffer, write_zip_buffer
 from sms_api.common.gateway.models import RouterConfig, ServerMode
 from sms_api.common.gateway.utils import get_local_simulation_outdir, get_simulation_outdir
 from sms_api.data.analysis_service import AnalysisService
-from sms_api.data.biocyc_service import BiocycService
-from sms_api.data.models import BiocycComponent, BiocycData
 from sms_api.data.parquet_service import ParquetService
 from sms_api.dependencies import (
     get_database_service,
@@ -515,30 +513,4 @@ async def download_example_file(
 
     except Exception as e:
         logger.exception("Error fetching the simulation analysis file.")
-        raise HTTPException(status_code=500, detail=str(e)) from e
-
-
-@config.router.get(
-    path="/download/ptools",
-    operation_id="download-ptools-data",
-    tags=["Data"],
-    # response_model=BiocycComponent,
-    summary="Download data for a given component from the Pathway Tools REST API",
-)
-async def download_ptools_data(
-    object_id: str = Query(
-        ..., example="--TRANS-ACENAPHTHENE-12-DIOL", description="Object ID of the component you wish to fetch"
-    ),
-    organism_id: str = Query(default="ECOLI"),
-    raw: bool = Query(
-        default=False,
-        description="If True, return an object containing both the BioCyc component and the request params/data",
-    ),
-) -> BiocycData | BiocycComponent:
-    try:
-        biocyc_svc = BiocycService()
-        biocyc_data = biocyc_svc.get_data(obj_id=object_id, org_id=organism_id)
-        return biocyc_data if raw else biocyc_data.to_dto()
-    except Exception as e:
-        logger.exception("Error.")
         raise HTTPException(status_code=500, detail=str(e)) from e
