+++ sms_api/version.py
@@ -1,2 +1,3 @@
-# STABLE: "0.2.8"
-__version__ = "0.2.11-dev"
+# LATEST STABLE: "0.2.11-dev"
+# PREVIOUS STABLE: "0.2.8"
+__version__ = "0.2.12-dev"
